refactor: move job submission fixtures from test file to conftest
